# openapi.yaml
openapi: 3.0.0
info:
  title: Microservices API
  version: 1.0.0
  description: A combined API for Comments, Users, and Posts services.

servers:
  - url: http://localhost
    description: Local Environment

paths:
  /users:
    get:
      summary: Get a list of users with pagination
      operationId: getUsers
      tags:
        - Users
      parameters:
        - name: page
          in: query
          description: The page number to retrieve
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: The number of users to return per page
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: A list of users with pagination information
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  hasMore:
                    type: boolean
                    description: Indicates whether there are more users to fetch
                    example: true
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid page or limit parameter"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Failed to fetch public profiles"
    post:
      summary: User registration
      operationId: registerUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                missingFields:
                  summary: Missing required fields
                  value:
                    message: "Email and password are required"
                invalidEmail:
                  summary: Invalid email format
                  value:
                    message: "Invalid email format"
                weakPassword:
                  summary: Weak password
                  value:
                    message: "Password must be at least 8 characters long"

  /users/login:
    post:
      summary: User login
      operationId: loginUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              description: Sets the access token as a cookie
              schema:
                $ref: "#/components/schemas/cookie"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
        "400":
          description: Invalid request

  /users/{id}/profile:
    get:
      summary: Get user profile
      operationId: getUserProfile
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid user ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid user ID"
        "404":
          description: User not found
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal server error"

  /users/{id}:
    put:
      summary: Update user details
      operationId: updateUser
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
                    example: "User details updated successfully"
        "400":
          description: Invalid user data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                missingFields:
                  summary: Missing required fields
                  value:
                    message: "Username and email are required"
                invalidEmail:
                  summary: Invalid email format
                  value:
                    message: "Invalid email format"
                weakPassword:
                  summary: Weak password
                  value:
                    message: "Password must be at least 8 characters long"            
                
        "404":
          description: User not foun
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User with given id not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal server error"

  /comments:
    post:
      summary: Add a comment
      operationId: addComment
      tags:
        - Comments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Comment"
      responses:
        "201":
          description: Comment created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  commentId:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
        "400":
          description: Invalid input
        "401":
          description: Unauthorized

  /comments/{commentId}:
    get:
      summary: Get a comment by ID
      operationId: getComment
      tags:
        - Comments
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comment found
        "404":
          description: Comment not found

  /posts/{postId}:
    get:
      summary: Get post details
      operationId: getPost
      tags:
        - Posts
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Post not found

components:
  schemas:
    CreateUserRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string

    UpdateUserRequest:
      type: object
      required:
        - email
        - username
      properties:
        email:
          type: string
        password:
          type: string
        username:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        bio:
          type: string
        profile_pic:
          type: string
          format: uri

    cookie:
      type: string
      example: access_token=abcde12345; Path=/; HttpOnly

    Comment:
      type: object
      required:
        - postId
        - userId
        - commentText
      properties:
        postId:
          type: string
        userId:
          type: string
        commentText:
          type: string
        parentCommentId:
          type: string
          nullable: true

    User:
      type: object
      required:
        - id
        - email
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
          nullable: true
        email:
          type: string
        status:
          type: string
          enum: ["active", "inactive", "banned"]
        role:
          type: string
          enum: ["user", "admin", "moderator"]
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        profile_pic:
          type: string
          nullable: true
          format: uri
        bio:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Post:
      type: object
      required:
        - postId
        - title
        - content
      properties:
        postId:
          type: string
        title:
          type: string
        content:
          type: string